add error classifiers
